Refactoring: moved SIMD mixer functions to separate code files; small speed-optimization by storing non-changing values to const class data members. Other cosmetic changes.
