+++ Mixer.cpp
@@ -2,11 +2,12 @@
 
 Mixer::Mixer(const Shared* const sh, const int n, const int m, const int s) : shared(sh),
   n(n), m(m), s(s), 
+  lowerLimitOfLearningRate(s == 1 ? MIN_LEARNING_RATE_S1 : MIN_LEARNING_RATE_SN),
+  isAdaptiveLearningRate(sh->GetOptionAdaptiveLearningRate()),
   scaleFactor(0), tx(n), wx(n * m), cxt(s), info(s), rates(s), pr(s) {
   for( uint64_t i = 0; i < s; ++i ) {
     pr[i] = 2048; //initial p=0.5
     rates[i] = MAX_LEARNING_RATE;
-    info[i].reset();
   }
   reset();
 }
