+++ Mixer.hpp
@@ -2,221 +2,6 @@
 
 #include "IPredictor.hpp"
 #include "Shared.hpp"
-#include "Utils.hpp"
-
-#if defined(__i386__) || defined(__x86_64__) || defined(_M_X64)
-#include <immintrin.h>
-#elif defined(__ARM_FEATURE_SIMD32) || defined(__ARM_NEON)
-#include <arm_neon.h>
-#endif
-
-#if (defined(__GNUC__) || defined(__clang__)) && (!defined(__ARM_FEATURE_SIMD32) && !defined(__ARM_NEON))
-__attribute__((target("avx512bw")))
-#endif
-static int dotProductSimdAvx512(const short* const t, const short* const w, int n) {
-#if !defined(__i386__) && !defined(__x86_64__) && !defined(_M_X64)
-  return 0;
-#else
-  __m512i sum = _mm512_setzero_si512();
-
-  while ((n -= 32) >= 0) {
-    __m512i tmp = _mm512_madd_epi16(*(__m512i*) & t[n], *(__m512i*) & w[n]);
-    tmp = _mm512_srai_epi32(tmp, 8);
-    sum = _mm512_add_epi32(sum, tmp);
-  }
-
-  __m256i lo = _mm512_extracti64x4_epi64(sum, 0);
-  __m256i hi = _mm512_extracti64x4_epi64(sum, 1);
-
-  __m256i newSum1 = _mm256_add_epi32(lo, hi);
-  __m128i newSum2 = _mm_add_epi32(_mm256_extractf128_si256(newSum1, 0), _mm256_extractf128_si256(newSum1, 1));
-  newSum2 = _mm_add_epi32(newSum2, _mm_srli_si128(newSum2, 8));
-  newSum2 = _mm_add_epi32(newSum2, _mm_srli_si128(newSum2, 4));
-  return _mm_cvtsi128_si32(newSum2);
-#endif
-}
-
-#if (defined(__GNUC__) || defined(__clang__)) && (!defined(__ARM_FEATURE_SIMD32) && !defined(__ARM_NEON))
-__attribute__((target("avx512bw")))
-#endif
-static void trainSimdAvx512(const short* const t, short* const w, int n, const int e) {
-#if !defined(__i386__) && !defined(__x86_64__) && !defined(_M_X64)
-  return;
-#else
-  const __m512i one = _mm512_set1_epi16(1);
-  const __m512i err = _mm512_set1_epi16(short(e));
-
-  while ((n -= 32) >= 0) {
-    __m512i tmp = _mm512_adds_epi16(*(__m512i*)&t[n], *(__m512i*)&t[n]);
-    tmp = _mm512_mulhi_epi16(tmp, err);
-    tmp = _mm512_adds_epi16(tmp, one);
-    tmp = _mm512_srai_epi16(tmp, 1);
-    tmp = _mm512_adds_epi16(tmp, *reinterpret_cast<__m512i*>(&w[n]));
-    *reinterpret_cast<__m512i*>(&w[n]) = tmp;
-  }
-#endif
-}
-
-#if (defined(__GNUC__) || defined(__clang__)) && (!defined(__ARM_FEATURE_SIMD32) && !defined(__ARM_NEON))
-__attribute__((target("avx2")))
-#endif
-static int dotProductSimdAvx2(const short *const t, const short *const w, int n) {
-#if !defined(__i386__) && !defined(__x86_64__) && !defined(_M_X64)
-  return 0;
-#else
-  __m256i sum = _mm256_setzero_si256();
-
-  while((n -= 16) >= 0 ) {
-    __m256i tmp = _mm256_madd_epi16(*(__m256i * ) & t[n], *(__m256i * ) & w[n]);
-    tmp = _mm256_srai_epi32(tmp, 8);
-    sum = _mm256_add_epi32(sum, tmp);
-  }
-
-  __m128i lo = _mm256_extractf128_si256(sum, 0);
-  __m128i hi = _mm256_extractf128_si256(sum, 1);
-
-  __m128i newSum = _mm_hadd_epi32(lo, hi);
-  newSum = _mm_add_epi32(newSum, _mm_srli_si128(newSum, 8));
-  newSum = _mm_add_epi32(newSum, _mm_srli_si128(newSum, 4));
-  return _mm_cvtsi128_si32(newSum);
-#endif
-}
-
-#if (defined(__GNUC__) || defined(__clang__)) && (!defined(__ARM_FEATURE_SIMD32) && !defined(__ARM_NEON))
-__attribute__((target("avx2")))
-#endif
-static void trainSimdAvx2(const short *const t, short *const w, int n, const int e) {
-#if !defined(__i386__) && !defined(__x86_64__) && !defined(_M_X64)
-  return;
-#else
-  const __m256i one = _mm256_set1_epi16(1);
-  const __m256i err = _mm256_set1_epi16(short(e));
-
-  while((n -= 16) >= 0 ) {
-    __m256i tmp = _mm256_adds_epi16(*(__m256i * ) & t[n], *(__m256i * ) & t[n]);
-    tmp = _mm256_mulhi_epi16(tmp, err);
-    tmp = _mm256_adds_epi16(tmp, one);
-    tmp = _mm256_srai_epi16(tmp, 1);
-    tmp = _mm256_adds_epi16(tmp, *reinterpret_cast<__m256i *>(&w[n]));
-    *reinterpret_cast<__m256i *>(&w[n]) = tmp;
-  }
-#endif
-}
-
-#if (defined(__GNUC__) || defined(__clang__)) && (defined(__ARM_FEATURE_SIMD32) || defined(__ARM_NEON))
-static inline int32x4_t _mm_mulhi_epi16(int32x4_t a, int32x4_t b){
-  int32x4_t rl = vmull_s16(vget_low_s16(vreinterpretq_s16_s32(a)), vget_low_s16(vreinterpretq_s16_s32(b)));
-  int32x4_t rh = vmull_s16(vget_high_s16(vreinterpretq_s16_s32(a)), vget_high_s16(vreinterpretq_s16_s32(b)));
-  uint16x8x2_t r = vuzpq_u16(vreinterpretq_u16_s32(rl), vreinterpretq_u16_s32(rh));
-  return vreinterpretq_s32_u16(r.val[1]);
-}
-
-static inline int32x4_t _mm_madd_epi16(int32x4_t a, int32x4_t b) {
-  int32x4_t pl = vmull_s16(vget_low_s16(vreinterpretq_s16_s32(a)), vget_low_s16(vreinterpretq_s16_s32(b)));
-  int32x4_t ph = vmull_s16(vget_high_s16(vreinterpretq_s16_s32(a)), vget_high_s16(vreinterpretq_s16_s32(b)));
-  int32x2_t rl = vpadd_s32(vget_low_s32(pl), vget_high_s32(pl));
-  int32x2_t rh = vpadd_s32(vget_low_s32(ph), vget_high_s32(ph));
-  return vcombine_s32(rl, rh);
-}
-#endif
-
-static int dotProductSimdNeon(const short *const t, const short *const w, int n) {
-#if (!defined(__ARM_FEATURE_SIMD32) && !defined(__ARM_NEON))
-  return 0;
-#else
-  int32x4_t sum = vdupq_n_s32(0);
-
-  while ((n -= 8) >= 0) {
-    int32x4_t tmp = _mm_madd_epi16(*(int32x4_t*) & t[n], *(int32x4_t*) & w[n]);
-    tmp = vshrq_n_s32(tmp, 8);
-    sum = vaddq_s32(sum, tmp);
-  }
-
-  sum = vaddq_s32(sum, vreinterpretq_s32_s8(vextq_s8(vreinterpretq_s8_s32(sum), vdupq_n_s8(0), 8)));
-  sum = vaddq_s32(sum, vreinterpretq_s32_s8(vextq_s8(vreinterpretq_s8_s32(sum), vdupq_n_s8(0), 4)));
-  return vgetq_lane_s32(sum, 0);
-#endif
-}
-
-static void trainSimdNeon(const short *const t, short *const w, int n, const int e) {
-#if (!defined(__ARM_FEATURE_SIMD32) && !defined(__ARM_NEON))
-  return;
-#else
-  const int32x4_t one = vreinterpretq_s32_s16(vdupq_n_s16(1));
-  const int32x4_t err = vreinterpretq_s32_s16(vdupq_n_s16(short(e)));
-
-  while ((n -= 8) >= 0) {
-    int32x4_t tmp = vreinterpretq_s32_s16(vqaddq_s16(vreinterpretq_s16_s32(*(int32x4_t*) & t[n]), vreinterpretq_s16_s32(*(int32x4_t*) & t[n])));
-    tmp = _mm_mulhi_epi16(tmp, err);
-    tmp = vreinterpretq_s32_s16(vqaddq_s16(vreinterpretq_s16_s32(tmp), vreinterpretq_s16_s32(one)));
-    tmp = vreinterpretq_s32_s16(vshrq_n_s16(vreinterpretq_s16_s32(tmp), (1)));
-    tmp = vreinterpretq_s32_s16(vqaddq_s16(vreinterpretq_s16_s32(tmp), vreinterpretq_s16_s32(*reinterpret_cast<int32x4_t*>(&w[n]))));
-    *reinterpret_cast<int32x4_t*>(&w[n]) = tmp;
-  }
-#endif
-}
-
-#if (defined(__GNUC__) || defined(__clang__)) && (!defined(__ARM_FEATURE_SIMD32) && !defined(__ARM_NEON))
-__attribute__((target("sse2")))
-#endif
-static int dotProductSimdSse2(const short *const t, const short *const w, int n) {
-#if !defined(__i386__) && !defined(__x86_64__) && !defined(_M_X64)
-  return 0;
-#else
-  __m128i sum = _mm_setzero_si128();
-
-  while((n -= 8) >= 0 ) {
-    __m128i tmp = _mm_madd_epi16(*(__m128i * ) & t[n], *(__m128i * ) & w[n]);
-    tmp = _mm_srai_epi32(tmp, 8);
-    sum = _mm_add_epi32(sum, tmp);
-  }
-
-  sum = _mm_add_epi32(sum, _mm_srli_si128(sum, 8));
-  sum = _mm_add_epi32(sum, _mm_srli_si128(sum, 4));
-  return _mm_cvtsi128_si32(sum);
-#endif
-}
-
-#if (defined(__GNUC__) || defined(__clang__)) && (!defined(__ARM_FEATURE_SIMD32) && !defined(__ARM_NEON))
-__attribute__((target("sse2")))
-#endif
-static void trainSimdSse2(const short *const t, short *const w, int n, const int e) {
-#if !defined(__i386__) && !defined(__x86_64__) && !defined(_M_X64)
-  return;
-#else
-  const __m128i one = _mm_set1_epi16(1);
-  const __m128i err = _mm_set1_epi16(short(e));
-
-  while((n -= 8) >= 0 ) {
-    __m128i tmp = _mm_adds_epi16(*(__m128i * ) & t[n], *(__m128i * ) & t[n]);
-    tmp = _mm_mulhi_epi16(tmp, err);
-    tmp = _mm_adds_epi16(tmp, one);
-    tmp = _mm_srai_epi16(tmp, 1);
-    tmp = _mm_adds_epi16(tmp, *reinterpret_cast<__m128i *>(&w[n]));
-    *reinterpret_cast<__m128i *>(&w[n]) = tmp;
-  }
-#endif
-}
-
-static int dotProductSimdNone(const short *const t, const short *const w, int n) {
-  int sum = 0;
-  while((n -= 2) >= 0 ) {
-    sum += (t[n] * w[n] + t[n + 1] * w[n + 1]) >> 8;
-  }
-  return sum;
-}
-
-static void trainSimdNone(const short *const t, short *const w, int n, const int err) {
-  while((n -= 1) >= 0 ) {
-    int wt = w[n] + ((((t[n] * err * 2) >> 16) + 1) >> 1);
-    if( wt < -32768 ) {
-      wt = -32768;
-    } else if( wt > 32767 ) {
-      wt = 32767;
-    }
-    *reinterpret_cast<short*>(&w[n]) = wt;
-  }
-}
 
 struct ErrorInfo {
   uint32_t data[2], sum, mask, collected;
@@ -227,18 +12,17 @@ struct ErrorInfo {
 };
 
 class Mixer : protected IPredictor {
-public:
-
+protected:
   static constexpr int MAX_LEARNING_RATE = 8 * 65536 - 1;
   static constexpr int MIN_LEARNING_RATE_S1 = 2 * 65536 - 1;
   static constexpr int MIN_LEARNING_RATE_SN = 6 * 65536 - 1;
-
-protected:
 
   const Shared * const shared;
   const uint32_t n; /**< max inputs */
   const uint32_t m; /**< max contexts */
   const uint32_t s; /**< max context sets */
+  const int lowerLimitOfLearningRate; /**< for linear learning rate decay */
+  const bool isAdaptiveLearningRate; /**< linked to command line option '-a' */
   int scaleFactor; /**< scale factor for dot product */
   Array<short, 64> tx; /**< n inputs from add() */
   Array<short, 64> wx; /**< n*m weights */
