+++ paq8px.vcxproj
@@ -315,6 +315,7 @@
     <ClCompile Include="ProgramChecker.cpp" />
     <ClCompile Include="Random.cpp" />
     <ClCompile Include="Shared.cpp" />
+    <ClCompile Include="SimdMixer.cpp" />
     <ClCompile Include="SmallStationaryContextMap.cpp" />
     <ClCompile Include="Squash.cpp" />
     <ClCompile Include="SSE.cpp" />
@@ -413,6 +414,12 @@
     <ClInclude Include="lstm\SimdLstmModel.hpp" />
     <ClInclude Include="Mixer.hpp" />
     <ClInclude Include="MixerFactory.hpp" />
+    <ClInclude Include="MixerFunctions_SIMD.hpp" />
+    <ClInclude Include="MixerFunctions_SIMD_AVX2.hpp" />
+    <ClInclude Include="MixerFunctions_SIMD_AVX512.hpp" />
+    <ClInclude Include="MixerFunctions_SIMD_Neon.hpp" />
+    <ClInclude Include="MixerFunctions_SIMD_None.hpp" />
+    <ClInclude Include="MixerFunctions_SIMD_SSE2.hpp" />
     <ClInclude Include="Models.hpp" />
     <ClInclude Include="model\Audio16BitModel.hpp" />
     <ClInclude Include="model\Audio8BitModel.hpp" />
