+++ paq8px.vcxproj.filters
@@ -315,6 +315,9 @@
     <ClCompile Include="StateMapPair.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="SimdMixer.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="AdaptiveMap.hpp">
@@ -769,6 +772,24 @@
     </ClInclude>
     <ClInclude Include="filter\tar.hpp">
       <Filter>filter</Filter>
+    </ClInclude>
+    <ClInclude Include="MixerFunctions_SIMD_AVX2.hpp">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="MixerFunctions_SIMD_AVX512.hpp">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="MixerFunctions_SIMD_Neon.hpp">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="MixerFunctions_SIMD_None.hpp">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="MixerFunctions_SIMD_SSE2.hpp">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="MixerFunctions_SIMD.hpp">
+      <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
